Restrict macro scanning to Cabbage section of .csd
The macro scanning implementation currently assumes the Cabbage section
is at the top of the .csd. Putting the Cabbage section at the bottom of
the .csd causes Cabbage to scan the entire .csd for macros instead of
just the Cabbage section. This change fixes macro scanning so it only
starts seeing macros after the "<Cabbage>" element is entered.