Free editor file/tab Csound resources before playback restarts
When playback on an editor file/tab is restarted, the Csound resources for the previous playback are not freed until after Csound is restarted for the new playback instance. This order of operations causes issues if the .csd being played back requires Csound to hold resources like OSC ports because the 2nd playback will fail since the first playback's Csound instance is still holding its resources.

This change fixes the issue by explicitly resetting any previous Csound instances when playback is restarted instead of waiting for the Juce framework to delete the graph node and plugin processor holding the previous Csound instance.