Fix setting of reserved Csound channels for operating system types
The reserved Csound channels for operating system types are not getting set correctly when types reported by the Juce framework are subtypes (e.g. `MacOSX_10_14` and `Windows10` instead of the top-level `MacOSX` and `Windows` types). This change fixes the issue by covering all current and future subtypes for `Linux`, `MacOSX` and `Windows` systems.