Keep file change timer running when Csound compile fails
When Csound fails to compile a file in the editor the timer that fires the callback to look for file changes is stopped and never restarted. Users then have to toggle the file tab `Stop/Play` button twice to get auto-loading from disk working again. This change keeps the file change watcher running when Csound compile fails.