Add --export-AU, --export-AUi, and --destination command line parameters
There are command line parameters for exporting VST and VSTi plugins, but there are no parameters for AU and AUi plugins and there is no way to set the destination path of the exported plugin. This change adds parameters for exporting AU and AUi plugins, and adds an optional `--destination` parameter for setting the destination path of the exported plugin.

This change also fixes an issue on MacOS causing .csd files set on the command line to be opened twice.