Don't open command lines with "--export-" when another instance starts
When an instance of Cabbage is running on MacOS and another instance is started with an --export-* command line parameter, Cabbage assumes the entire command line is a .csd file. When the .csd file is not found Cabbage shows the file open dialog. This change fixes the issue by only opening the command line if "--export-" is not found in it.